Mapper, XML 커밋_2
