--- src/main/java/com/example/lit/domain/vo/LikeVO.java
@@ -1,11 +1,0 @@
-package com.example.lit.domain.vo;
-
-import lombok.Data;
-import org.springframework.stereotype.Component;
-
-@Component
-@Data
-public class LikeVO {
-    private Long reviewNumber;
-    private Long userNumber;
-}
