--- src/main/java/com/example/lit/domain/vo/MedalDTO.java
@@ -1,4 +1,0 @@
-package com.example.lit.domain.vo;
-
-public class MedalDTO {
-}
