+++ src/main/java/com/example/lit/domain/vo/project/ProjectFileVO.java
@@ -1,4 +1,4 @@
-package com.example.lit.domain.vo;
+package com.example.lit.domain.vo.project;
 
 import lombok.Data;
 import org.springframework.stereotype.Component;
