+++ src/main/java/com/example/lit/domain/vo/user/achievement/MedalDTO.java
@@ -1,0 +1,4 @@
+package com.example.lit.domain.vo.user.achievement;
+
+public class MedalDTO {
+}
