+++ src/main/java/com/example/lit/mapper/project/ProjectFileMapper.java
@@ -1,7 +1,6 @@
 package com.example.lit.mapper.project;
 
-import com.example.lit.domain.vo.ProjectFileVO;
-import com.example.lit.domain.vo.ReviewFileVO;
+import com.example.lit.domain.vo.project.ProjectFileVO;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
