--- src/main/java/com/example/lit/mapper/user/AchievementMapper.java
@@ -1,7 +1,0 @@
-package com.example.lit.mapper.user;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface AchievementMapper {
-}
