+++ src/main/resources/config/config.xml
@@ -2,5 +2,19 @@
 <!DOCTYPE configuration PUBLIC "-//mybatis.org//DTD Config 3.0/EN" "http://mybatis.org/dtd/mybatis-3-config.dtd">
 <configuration>
     <typeAliases>
+        <typeAlias type="com.example.lit.domain.vo.project.ParticipationVO" alias="participationVO"/>
+        <typeAlias type="com.example.lit.domain.vo.project.ProjectFileVO" alias="projectFileVO"/>
+        <typeAlias type="com.example.lit.domain.vo.project.ProjectVO" alias="projectVO"/>
+        <typeAlias type="com.example.lit.domain.vo.review.LikeVO" alias="likeVO"/>
+        <typeAlias type="com.example.lit.domain.vo.review.ReplyVO" alias="replyVO"/>
+        <typeAlias type="com.example.lit.domain.vo.review.ReportVO" alias="reportVO"/>
+        <typeAlias type="com.example.lit.domain.vo.review.ReviewFileVO" alias="reviewFileVO"/>
+        <typeAlias type="com.example.lit.domain.vo.review.ReviewVO" alias="reviewVO"/>
+        <typeAlias type="com.example.lit.domain.vo.user.FollowVO" alias="followVO"/>
+        <typeAlias type="com.example.lit.domain.vo.user.MessageVO" alias="messageVO"/>
+        <typeAlias type="com.example.lit.domain.vo.user.UserFileVO" alias="userFileVO"/>
+        <typeAlias type="com.example.lit.domain.vo.user.UserVO" alias="userVO"/>
+        <typeAlias type="com.example.lit.domain.vo.user.achievement.AchievementVO" alias="achievementVO"/>
+        <typeAlias type="com.example.lit.domain.vo.user.achievement.MedalDTO" alias="medalDTO"/>
     </typeAliases>
 </configuration>
